agentic-ai-usecases/advanced/flight-booking-assistant/requirements.txt
langgraph>=1.0.7
openai>=2.16.0
streamlit>=1.50.0
python-dotenv